Handle keyboard submit with a decorator instead of a prop
