--- src/useKeyboardSave.ts
@@ -1,26 +1,0 @@
-import * as React from 'react'
-
-export const useKeyboardSave = (
-  callback?: (
-    event?: Partial<
-      Pick<React.SyntheticEvent, 'preventDefault' | 'stopPropagation'>
-    >
-  ) => void
-) => {
-  React.useEffect(() => {
-    if (callback) {
-      const handleKeyDown = (e: KeyboardEvent) => {
-        if ((e.ctrlKey || e.metaKey) && e.key === 's') {
-          e.preventDefault()
-          callback(e)
-        }
-      }
-
-      document.addEventListener('keydown', handleKeyDown)
-
-      return () => {
-        document.removeEventListener('keydown', handleKeyDown)
-      }
-    }
-  }, [callback])
-}
